+++ samples/RustPlus.ConsoleApp/Features/SetSubscription.cs
@@ -5,7 +5,7 @@ namespace RustPlus.ConsoleApp.Features;
 
 internal sealed class SetSubscription(IRustPlus rustPlus)
 {
-    public async Task SetSubscriptionAsync(uint entityId, bool doSubscribe)
+    public async Task SetSubscriptionAsync(ulong entityId, bool doSubscribe)
     {
         var response = await rustPlus.SetSubscriptionAsync(entityId, doSubscribe);
         DisplayUtilities.DisplayJson("SetSubscription", response);
